Renaming snapshot-test to snapshot-kafka-test
